scrapy_news_collector/requirements.txt
scrapy==2.11.0
scrapy-user-agents==0.1.1
requests==2.31.0
lxml==4.9.3
feedparser==6.0.10
openai==1.35.0
python-dotenv==1.0.0
python-dateutil==2.8.2